# feat: optimized Digital Ocean P2P testing with performance tiers and timing analysis